refactor: move payment-related models and functionality to the payments app
chore: remove some unused codes
