+++ config/settings/base.py
@@ -70,6 +70,7 @@ LOCAL_APPS = [
     "sbily.users",
     "sbily.authentication",
     "sbily.notifications",
+    "sbily.payments",
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
