+++ sbily/payments/templates/confirm_payment.html
