+++ sbily/users/views.py
@@ -8,23 +8,17 @@ from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.http import HttpRequest
-from django.http import HttpResponse
 from django.http import JsonResponse
 from django.shortcuts import redirect
 from django.shortcuts import render
-from django.urls import reverse
-from django.utils.timezone import now
-from django.utils.timezone import timedelta
-from django.views.decorators.csrf import csrf_exempt
 
 from sbily.utils.data import validate
 from sbily.utils.data import validate_password
 from sbily.utils.errors import BadRequestError
 from sbily.utils.errors import bad_request_error
-from sbily.utils.urls import redirect_with_params
+from sbily.utils.urls import redirect_with_tab
 
 from .forms import ProfileForm
-from .models import Subscription
 from .models import Token
 from .models import User
 from .tasks import send_deleted_account_email
@@ -32,16 +26,10 @@ from .tasks import send_email_change_instructions
 from .tasks import send_email_changed_email
 from .tasks import send_email_verification
 from .tasks import send_password_changed_email
-from .webhook import handle_stripe_webhook
 
 stripe.api_key = settings.STRIPE_SECRET_KEY
 
 logger = logging.getLogger("users.views")
-
-
-def redirect_with_tab(tab: str, **kwargs):
-    params = {"tab": tab, **kwargs}
-    return redirect_with_params("my_account", params)
 
 
 @login_required
@@ -246,7 +234,14 @@ def delete_account(request: HttpRequest):
         user.delete()
 
         if customer:
-            customer.delete()
+            try:
+                customer.delete()
+            except stripe.error.StripeError as e:
+                logger.warning(
+                    "StripeError encountered while deleting customer's account for user %s: %s",  # noqa: E501
+                    user.email,
+                    e,
+                )
 
         messages.success(request, "Account deleted successfully")
         return redirect("sign_in")
@@ -262,341 +257,3 @@ def set_user_timezone(request: HttpRequest):
     if "timezone" in request.POST:
         request.session["user_timezone"] = request.POST["timezone"]
     return JsonResponse({"status": "ok"})
-
-
-@login_required
-def upgrade_plan(request: HttpRequest):
-    if request.method != "GET":
-        return redirect_with_tab("plan")
-
-    try:
-        user = request.user
-        if user.is_premium:
-            bad_request_error("You are already a premium user")
-
-        customer = user.get_stripe_customer()
-        setup_intent = stripe.SetupIntent.create(
-            customer=customer.id,
-            payment_method_types=["card"],
-        )
-
-        return render(
-            request,
-            "payment/setup_card.html",
-            {
-                "client_secret": setup_intent.client_secret,
-                "redirect_url": request.build_absolute_uri(
-                    reverse("finalize_upgrade"),
-                ),
-            },
-        )
-
-    except BadRequestError as e:
-        messages.error(request, e.message)
-        return redirect_with_tab("plan")
-    except stripe.error.StripeError as e:
-        messages.error(request, f"Payment error: {e!s}")
-        return redirect_with_tab("plan")
-
-
-@login_required
-def finalize_upgrade(request: HttpRequest):
-    """Handle the redirect after card setup for upgrade"""
-    if request.method != "GET":
-        return redirect_with_tab("plan")
-
-    payment_method = request.GET.get("payment_method")
-    setup_intent = request.GET.get("setup_intent")
-
-    if not validate([payment_method, setup_intent]):
-        messages.error(request, "Missing payment information")
-        return redirect_with_tab("plan")
-
-    try:
-        user = request.user
-        if user.is_premium:
-            bad_request_error("You are already a premium user")
-
-        subscription, _ = Subscription.objects.get_or_create(
-            user=user,
-            defaults={
-                "price": 5.00,  # Monthly premium price
-                "status": Subscription.STATUS_INCOMPLETE,
-                "start_date": now(),
-                "end_date": now() + timedelta(days=30),
-                "is_auto_renew": True,
-            },
-        )
-
-        # Create Stripe subscription
-        result = subscription.create_stripe_subscription(
-            payment_method_id=payment_method,
-        )
-
-        if result["status"] == "success":
-            user.upgrade_to_premium()
-            messages.success(request, "Successfully upgraded to premium!")
-            return redirect_with_tab("plan")
-        if result["status"] == "action_required":
-            return render(
-                request,
-                "payment/confirm_payment.html",
-                {
-                    "client_secret": result["client_secret"],
-                    "redirect_url": request.build_absolute_uri(
-                        reverse("subscription_complete"),
-                    ),
-                },
-            )
-        subscription.delete()  # Clean up failed subscription
-        messages.error(
-            request,
-            f"Payment failed: {result.get('error', 'Unknown error')}",
-        )
-        return redirect_with_tab("plan")
-
-    except Exception as e:
-        messages.error(request, f"Error processing upgrade: {e!s}")
-        return redirect_with_tab("plan")
-
-
-@login_required
-def subscription_complete(request: HttpRequest):
-    """Handle completion of subscription payment"""
-    if request.method != "GET":
-        return redirect_with_tab("plan")
-
-    payment_intent = request.GET.get("payment_intent")
-
-    if not payment_intent:
-        messages.error(request, "Missing payment information")
-        return redirect_with_tab("plan")
-
-    try:
-        intent = stripe.PaymentIntent.retrieve(payment_intent)
-
-        if intent.status == "succeeded":
-            user = request.user
-            user.upgrade_to_premium()
-            messages.success(request, "Successfully upgraded to premium!")
-        else:
-            messages.error(request, f"Payment not completed: {intent.status}")
-
-        return redirect_with_tab("plan")
-    except stripe.error.StripeError as e:
-        messages.error(request, f"Payment verification error: {e!s}")
-        return redirect_with_tab("plan")
-
-
-@login_required
-def downgrade_plan(request: HttpRequest):
-    if request.method != "POST":
-        return redirect_with_tab("plan")
-
-    try:
-        user = request.user
-        if not user.is_premium:
-            bad_request_error("You are not a premium user")
-
-        subscription = Subscription.objects.filter(
-            user=user,
-            status=Subscription.STATUS_ACTIVE,
-        ).first()
-
-        if subscription and subscription.stripe_subscription_id:
-            result = subscription.cancel_stripe_subscription()
-            if result["status"] != "success":
-                messages.warning(
-                    request,
-                    f"Warning: Stripe cancellation issue - {result.get('error')}",
-                )
-
-        # Proceed with downgrade anyway
-        if message := user.downgrade_to_free():
-            messages.warning(request, message)
-
-        messages.success(request, "Successfully downgraded to free!")
-        return redirect_with_tab("plan")
-    except BadRequestError as e:
-        messages.error(request, e.message)
-        return redirect_with_tab("plan")
-    except Exception as e:
-        messages.error(request, f"Error downgrading to free: {e}")
-        return redirect_with_tab("plan")
-
-
-@login_required
-def purchase_links(request: HttpRequest):
-    if request.method != "POST":
-        return redirect_with_tab("plan")
-
-    try:
-        user = request.user
-        if not user.is_premium:
-            bad_request_error(
-                "You need to be a premium user to purchase additional links",
-            )
-
-        customer = user.get_stripe_customer()
-        if not customer.invoice_settings.default_payment_method:
-            messages.error(request, "Please add a payment method first")
-            return add_payment_method(request)
-
-        link_type = request.POST.get("link_type")
-        quantity = int(request.POST.get("quantity", 0))
-
-        if not link_type or quantity <= 0:
-            bad_request_error("Invalid link type or quantity")
-
-        if link_type not in ["permanent", "temporary"]:
-            bad_request_error("Invalid link type")
-
-        result = user.process_package_payment(
-            link_type=link_type,
-            quantity=quantity,
-        )
-
-        if result["status"] == "success":
-            messages.success(
-                request,
-                f"Successfully purchased {quantity} {link_type} links!",
-            )
-        elif result["status"] == "action_required":
-            return render(
-                request,
-                "payment/confirm_payment.html",
-                {
-                    "client_secret": result["client_secret"],
-                    "redirect_url": request.build_absolute_uri(
-                        reverse("purchase_complete"),
-                    ),
-                },
-            )
-        else:
-            messages.error(
-                request,
-                f"Payment failed: {result.get('error', 'Unknown error')}",
-            )
-    except BadRequestError as e:
-        messages.error(request, e.message)
-    except Exception as e:
-        messages.error(request, f"Error purchasing links: {e!s}")
-
-    return redirect_with_tab("plan")
-
-
-@login_required
-def add_payment_method(request: HttpRequest):
-    """Add a new payment method to the user's account"""
-    try:
-        user = request.user
-
-        customer = user.get_stripe_customer()
-        setup_intent = stripe.SetupIntent.create(
-            customer=customer.id,
-            payment_method_types=["card"],
-        )
-
-        return render(
-            request,
-            "payment/setup_card.html",
-            {
-                "client_secret": setup_intent.client_secret,
-                "redirect_url": request.build_absolute_uri(
-                    reverse("payment_method_added"),
-                ),
-            },
-        )
-    except Exception as e:
-        messages.error(request, f"Error setting up payment method: {e!s}")
-        return redirect_with_tab("plan")
-
-
-@login_required
-def payment_method_added(request: HttpRequest):
-    """Handle redirect after adding payment method"""
-    if payment_method := request.GET.get("payment_method"):
-        try:
-            user = request.user
-
-            # Attach payment method to customer
-            customer = user.get_stripe_customer()
-            stripe.PaymentMethod.attach(
-                payment_method,
-                customer=customer.id,
-            )
-
-            stripe.Customer.modify(
-                customer.id,
-                invoice_settings={
-                    "default_payment_method": payment_method,
-                },
-            )
-
-            user.update_card_details(payment_method)
-            messages.success(request, "Payment method successfully added!")
-        except Exception as e:
-            messages.error(request, f"Error adding payment method: {e!s}")
-
-    return redirect_with_tab("plan")
-
-
-@login_required
-def purchase_complete(request: HttpRequest):
-    """Handle completion of link purchase payment"""
-    if request.method != "GET":
-        return redirect_with_tab("plan")
-
-    payment_intent = request.GET.get("payment_intent")
-
-    if not payment_intent:
-        messages.error(request, "Missing payment information")
-        return redirect_with_tab("plan")
-
-    try:
-        intent = stripe.PaymentIntent.retrieve(payment_intent)
-
-        if intent.status == "succeeded":
-            metadata = intent.metadata
-            link_type = metadata.get("package_type")
-            quantity = int(metadata.get("quantity", 0))
-
-            if link_type and quantity > 0:
-                messages.success(
-                    request,
-                    f"Successfully purchased {quantity} {link_type} links!",
-                )
-            else:
-                messages.success(request, "Payment completed successfully!")
-        else:
-            messages.error(request, f"Payment not completed: {intent.status}")
-
-        return redirect_with_tab("plan")
-    except stripe.error.StripeError as e:
-        messages.error(request, f"Payment verification error: {e!s}")
-        return redirect_with_tab("plan")
-
-
-@csrf_exempt
-def stripe_webhook(request):
-    """Handle Stripe webhook events"""
-    payload = request.body
-    sig_header = request.headers.get("stripe-signature")
-
-    try:
-        event = stripe.Webhook.construct_event(
-            payload,
-            sig_header,
-            settings.STRIPE_WEBHOOK_SECRET,
-        )
-    except ValueError:
-        return HttpResponse(status=400)
-    except stripe.error.SignatureVerificationError:
-        return HttpResponse(status=400)
-
-    try:
-        handle_stripe_webhook(event)
-        return JsonResponse({"status": "success"})
-    except Exception as e:
-        logger.exception("Error handling Stripe webhook", exc_info=e)
-        return JsonResponse({"status": "error", "message": str(e)}, status=500)
